refactor: rename the modules to make sure tree-shaking works
