+++ components/css-baseline-deprecated/index.ts
@@ -1,0 +1,6 @@
+import CssBaseline from '../css-baseline'
+import useWarning from '../utils/use-warning'
+
+useWarning('Module "CSSBaseline" is deprecated. Use `CssBaseline` instead of it.', 'CSSBaseline')
+
+export default CssBaseline
