+++ components/index.ts
@@ -3,12 +3,14 @@
 export * from './styles/themes'
 export { default as Utils } from './utils-shared/index'
 export { default as useTheme } from './styles/use-theme'
-export { default as ZEITUIProvider } from './providers/zeit-ui-provider'
+export { default as ZEITUIProvider } from './zeit-provider-deprecated'
+export { default as ZeitProvider } from './zeit-provider'
+export { default as CSSBaseline } from './css-baseline-deprecated'
+export { default as CssBaseline } from './css-baseline'
 export { default as useToasts } from './toast'
 export { default as useInput } from './input/use-input'
 export { default as useModal } from './modal/use-modal'
 export { default as useTabs } from './tabs/use-tabs'
-export { default as CSSBaseline } from './styles/css-baseline'
 export { default as Avatar } from './avatar'
 export { default as Text } from './text'
 export { default as Note } from './note'
