+++ components/zeit-provider-deprecated/index.ts
@@ -1,0 +1,9 @@
+import ZeitProvider from '../zeit-provider'
+import useWarning from '../utils/use-warning'
+
+useWarning(
+  'Module "ZEITUIProvider" is deprecated. Use `ZeitProvider` instead of it.',
+  'ZEITUIProvider',
+)
+
+export default ZeitProvider
