+++ components/zeit-provider/index.ts
@@ -1,0 +1,3 @@
+import ZeitProvider from './zeit-provider'
+
+export default ZeitProvider
